Update on LogInController
- "Aceptar" button is enable/disable when it should
- Email validation on "Aceptar" button
- Opens MainWindow
