+++ Reto1Library/src/libraries/MessageType.java
@@ -13,7 +13,6 @@ package libraries;
 public enum MessageType {
     LogIn,
     SignUp,
-    Ex_ClassNotFound,
     Ex_Credentials,
     Ex_EmailAlreadyExists,
     Ex_ServerError
