+++ RetoSignInCliente/nbproject/project.properties
@@ -29,23 +29,20 @@ dist.jar=${dist.dir}/RetoSignInCliente.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 endorsed.classpath=
 excludes=
-file.reference.NetBeansProjects-Reto1Library=C:\\Users\\2dam\\Documents\\NetBeansProjects\\Reto1Library
-file.reference.Reto1Library.jar=C:\\Users\\2dam\\Documents\\NetBeansProjects\\Reto1Library\\dist\\Reto1Library.jar
+file.reference.Reto1Library.jar-1=..\\Reto1Library\\dist\\Reto1Library.jar
 includes=**
 # Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
 jar.archive.disabled=true
 jar.compress=false
 javac.classpath=\
     ${javafx.classpath.extension}:\
-    ${file.reference.NetBeansProjects-Reto1Library}:\
-    ${file.reference.Reto1Library.jar}
+    ${file.reference.Reto1Library.jar-1}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
 javac.external.vm=false
 javac.processorpath=\
-    ${javac.classpath}:\
-    ${file.reference.Reto1Library.jar}
+    ${javac.classpath}
 javac.source=1.8
 javac.target=1.8
 javac.test.classpath=\
@@ -109,7 +106,6 @@ mkdist.disabled=false
 platform.active=default_platform
 run.classpath=\
     ${javac.classpath}:\
-    ${file.reference.Reto1Library.jar}:\
     ${build.classes.dir}
 run.test.classpath=\
     ${javac.test.classpath}:\
