+++ RetoSignInCliente/src/controller/LogInController.java
@@ -5,8 +5,11 @@
  */
 package controller;
 
+import exceptions.EmailFormatException;
 import factory.SignableFactory;
 import java.io.IOException;
+import java.util.regex.Pattern;
+import javafx.beans.Observable;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
@@ -20,6 +23,8 @@ import javafx.scene.control.TextField;
 import javafx.scene.input.InputMethodEvent;
 import javafx.stage.Stage;
 import libraries.User;
+import javafx.beans.value.ChangeListener;
+import javafx.beans.value.ObservableValue;
 
 /**
  * This class contains the responses for behavior of the Log In view
@@ -31,7 +36,7 @@ public class LogInController {
     private Stage stage;
 
     @FXML
-    private Label lblEmail, lblPassword;
+    private Label lblEmail, lblPassword, lblError;
     @FXML
     private TextField txtEmail;
     @FXML
@@ -50,7 +55,7 @@ public class LogInController {
         try {
             Scene scene = new Scene(root);
             stage.setScene(scene);
-//TO-DO: mirar que se cargue la escena bien
+            //TO-DO: mirar que se cargue la escena bien
             // Establish window title
             stage.setTitle("Iniciar sesión");
             // Window is not resizable
@@ -70,9 +75,6 @@ public class LogInController {
             hrefSignUp.setOnAction(this::handleHrefSignupAction);
             txtEmail.textProperty().addListener(this::handleTextChange);
             pwdPassword.textProperty().addListener(this::handleTextChange);
-            //txtEmail.addEventHandler(InputMethodEvent.INPUT_METHOD_TEXT_CHANGED, this::handleTextChange);
-            //pwdPassword.addEventHandler(InputMethodEvent.INPUT_METHOD_TEXT_CHANGED, this::handleTextChange);
-
         } catch (Exception ex) {
             ex.getMessage();
         }
@@ -95,34 +97,45 @@ public class LogInController {
             String email = txtEmail.getText();
             String password = pwdPassword.getText();
 
+            //Validate the format of the email, it must have a text before an '@' and a text before and after '.'
+            //Pattern that must be respected
+            String regexEmail = "^[A-Za-z0-9]+@[A-Za-z0-9]+\\.[A-Za-z]{2,}$";
+            Pattern patternEmail = Pattern.compile(regexEmail);
+            //Validate if the pattern doesn't match the txtEmail field text
+            if (!patternEmail.matcher(txtEmail.getText()).matches()) {
+                throw new EmailFormatException("This email Format is not correct");
+            }
+
             // Create a User object with the provided data
             User user = new User();
             user.setLogin(email);
             user.setPassword(password);
 
             // Send the User created to the logic Tier and recieve a full informed User
-            User mainWindowUser = new User();
-            mainWindowUser = SignableFactory.getSignable().logIn(user);
+            //User mainWindowUser = new User();
+            //mainWindowUser = SignableFactory.getSignable().logIn(user);
 
             // Open a MainWindow instance and close this window
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/view/MainWindow.fxml"));
-            Parent root = loader.load();
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/view/MainWindowFXML.fxml"));
+            Parent root = (Parent) loader.load();
             MainWindowController mainWindowController = loader.getController();
-            mainWindowController.initStage(root, mainWindowUser);
+            mainWindowController.initStage(root, user);
 
             stage.close();
 
-        } catch (Exception ex) { //de momento está puesta una exception generica pero aqui deberian saltar las excepciones de conexion a la base de datos y de credenciales
-            ex.getMessage();
+        } catch (EmailFormatException ex) { //de momento está puesta una exception generica pero aqui deberian saltar las excepciones de conexion a la base de datos y de credenciales
+            showError(ex.getMessage());
+        } catch (Exception ex){
+            ex.printStackTrace();
         }
     }
 
     @FXML
-    private void handleTextChange(InputMethodEvent e) {
+    private void handleTextChange(Observable observable) {
         try {
             String email = txtEmail.getText();
             String password = pwdPassword.getText();
-
+            /*
             if (!email.trim().isEmpty() && !password.trim().isEmpty()) {
                 // Enable 'Entrar' button
                 loginButton.setDisable(false);
@@ -130,6 +143,15 @@ public class LogInController {
             } else {
                 // Disable 'Entrar' button
                 loginButton.setDisable(true);
+            }
+             */
+            if (email.trim().isEmpty() || password.trim().isEmpty()) {
+                // Disable 'Entrar' button
+                loginButton.setDisable(true);
+
+            } else {
+                // Enable 'Entrar' button
+                loginButton.setDisable(false);
             }
 
         } catch (Exception ex) {
@@ -152,4 +174,10 @@ public class LogInController {
             // Gestionar la exception
         }
     }
+    
+    private void showError(String e){
+        //Show error message
+        lblError.setText(e);
+    }
+    
 }
