+++ RetoSignInCliente/src/view/LogInFXML.fxml
@@ -13,6 +13,7 @@
       <Label fx:id="lblEmail" layoutX="199.0" layoutY="146.0" prefHeight="18.0" prefWidth="48.0" text="Email" />
       <Label fx:id="lblPassword" layoutX="199.0" layoutY="203.0" text="Contraseña" />
       <TextField id="Email" fx:id="txtEmail" layoutX="199.0" layoutY="164.0" prefHeight="25.0" prefWidth="201.0" />
+      <PasswordField fx:id="pwdPassword" layoutX="199.0" layoutY="227.0" prefHeight="25.0" prefWidth="201.0" />
       <Hyperlink fx:id="hrefSignUp" layoutX="193.0" layoutY="294.0" text="¿No tienes usuario? Registrate Aquí" />
 
       <Label layoutX="251.0" layoutY="69.0" prefHeight="18.0" prefWidth="104.0" text="Iniciar Sesión" textAlignment="CENTER">
@@ -21,6 +22,6 @@
          </font>
       </Label>
       <Button fx:id="loginButton" defaultButton="true" layoutX="532.0" layoutY="354.0" mnemonicParsing="false" onAction="#handleLoginButtonAction" text="Entrar" />
-      <PasswordField fx:id="pwdPassword" layoutX="199.0" layoutY="227.0" prefHeight="25.0" prefWidth="201.0" />
+      <Label id="lblError" fx:id="lblError" layoutX="408.0" layoutY="168.0" textFill="RED" />
    </children>
 </Pane>
